+++ apps/web/next.config.js
@@ -3,4 +3,5 @@ module.exports = {
   output: "export",
   distDir: "../../packages/cli/dist/webapp",
   productionBrowserSourceMaps: true,
+  optimizeFonts: false,
 };