+++ packages/cli/src/bin/index.ts
@@ -14,7 +14,13 @@ import { loadDataset } from "./dataset";
 import { DatasetError } from "../error";
 import { DefaultRunsConfigType, getDefaultRunsConfig } from "../runs";
 import { Dataset, RunCompletion } from "@empiricalrun/types";
-import { printStatsSummary, setRunSummary } from "../stats";
+import {
+  failedOutputsSummary,
+  printStatsSummary,
+  setRunSummary,
+} from "../stats";
+import { reportOnCI } from "../reporters/ci";
+import detect from "detect-port";
 
 const configFileName = "empiricalrc.json";
 const cwd = process.cwd();
@@ -70,7 +76,7 @@ program
     } catch (err) {
       console.log(`${red("[Error]")} Failed to read ${configFileName} file`);
       console.log(yellow("Please ensure running init command first"));
-      return;
+      process.exit(1);
     }
 
     console.log(`${green("[Success]")} - read ${configFileName} file`);
@@ -84,7 +90,7 @@ program
     } catch (error) {
       if (error instanceof DatasetError) {
         console.log(`${red("[Error]")} ${error.message}`);
-        return;
+        process.exit(1);
       } else {
         throw error;
       }
@@ -95,9 +101,8 @@ program
     );
     const completion = await Promise.all(
       runs.map((r) => {
-        if (r.type === "py-script") {
-          r.pythonPath = options.pythonPath;
-        }
+        r.parameters = r.parameters ? r.parameters : {};
+        r.parameters.pythonPath = options.pythonPath;
         return execute(r, dataset, () => {
           progressBar.increment();
         });
@@ -108,6 +113,7 @@ program
     setRunSummary(completion);
     printStatsSummary(completion);
 
+    // TODO: this is not sent in CI report
     console.log(bold("Total dataset samples:"), dataset.samples?.length || 0);
     const endTime = performance.now();
     console.log(
@@ -123,20 +129,47 @@ program
       };
       await fs.mkdir(`${cwd}/${cacheDir}`, { recursive: true });
       await fs.writeFile(outputFilePath, JSON.stringify(data, null, 2));
+    } else {
+      await reportOnCI(completion, dataset);
+    }
+
+    const failedOutputs = failedOutputsSummary(completion);
+    if (failedOutputs) {
+      const { code, message } = failedOutputs;
+      console.log(
+        `${red("[Error]")} Some outputs were not generated successfully`,
+      );
+      console.log(`${red("[Error]")} ${code}: ${message}`);
+      process.exit(1);
     }
   });
 
+const defaultWebUIPort = 1337;
 program
   .command("ui")
   .description("visualise the results of a run in your web browser")
-  .action(async () => {
+  .option(
+    "-p, --port <int>",
+    "port to run the empirical webapp on",
+    `${defaultWebUIPort}`,
+  )
+  .action(async (options) => {
     console.log(yellow("Initiating webapp..."));
     const app = express();
-    const port = 8000;
+    const port =
+      !options.port || isNaN(Number(options.port))
+        ? defaultWebUIPort
+        : Number(options.port);
+    const availablePort = await detect(port);
+    if (availablePort !== port) {
+      console.log(
+        `${yellow("[Warning]")} Port ${port} is unavailable. Trying port ${availablePort}.`,
+      );
+    }
     app.use(express.static(path.join(__dirname, "../webapp")));
     app.get("/api/results", (req, res) => res.sendFile(outputFilePath));
-    const fullUrl = `http://localhost:${port}`;
-    app.listen(port, () => {
+    const fullUrl = `http://localhost:${availablePort}`;
+    app.listen(availablePort, () => {
       console.log(`Empirical app running on ${fullUrl}`);
       opener(fullUrl);
     });