+++ packages/core/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @empiricalrun/core
 
+## 0.2.0
+
+### Minor Changes
+
+- 90082c8: feat: support model parameters with passthrough
+- 17860b0: feat: add support for passthrough config for script executor
+
+### Patch Changes
+
+- b354507: fix: change script value property to path in run
+- Updated dependencies [b354507]
+- Updated dependencies [90082c8]
+- Updated dependencies [a9e840a]
+- Updated dependencies [b3cb527]
+- Updated dependencies [d7f4b53]
+- Updated dependencies [a4cfb49]
+  - @empiricalrun/scorer@0.1.3
+  - @empiricalrun/ai@0.2.0
+
+## 0.1.3
+
+### Patch Changes
+
+- 571986a: chore: update npm package license
+- Updated dependencies [571986a]
+- Updated dependencies [4a858f5]
+  - @empiricalrun/scorer@0.1.2
+  - @empiricalrun/ai@0.1.1
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [424f994]
+  - @empiricalrun/scorer@0.1.1
+
 ## 0.1.1
 
 ### Patch Changes