+++ packages/types/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @empiricalrun/types
 
+## 0.2.0
+
+### Minor Changes
+
+- 90082c8: feat: support model parameters with passthrough
+- 17860b0: feat: add support for passthrough config for script executor
+
+### Patch Changes
+
+- b354507: fix: change script value property to path in run
+
+## 0.1.1
+
+### Patch Changes
+
+- 571986a: chore: update npm package license
+
 ## 0.1.0
 
 ### Minor Changes